crypto_alert_bot/requirements.txt
aiogram==3.7.0
aiosqlite==0.20.0
aiohttp==3.9.5
python-dotenv==1.0.1
cachetools==5.3.3
